Move some util functions to the api package.
