+++ api/types/github_test.go
@@ -19,8 +19,6 @@ package types
 import (
 	"testing"
 
-	"github.com/gravitational/teleport/lib/utils"
-
 	check "gopkg.in/check.v1"
 )
 
@@ -29,7 +27,7 @@ type GithubSuite struct{}
 var _ = check.Suite(&GithubSuite{})
 
 func (s *GithubSuite) SetUpSuite(c *check.C) {
-	utils.InitLoggerForTests(testing.Verbose())
+	InitLoggerForTests(testing.Verbose())
 }
 
 func (s *GithubSuite) TestUnmarshal(c *check.C) {
