+++ api/types/namespace.go
@@ -21,8 +21,6 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/gravitational/teleport/lib/utils"
-
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
 )
@@ -150,7 +148,7 @@ func MarshalNamespace(resource Namespace, opts ...MarshalOption) ([]byte, error)
 		copy.SetResourceID(0)
 		resource = copy
 	}
-	return utils.FastMarshal(resource)
+	return FastMarshal(resource)
 }
 
 // UnmarshalNamespace unmarshals role from JSON or YAML,
@@ -168,7 +166,7 @@ func UnmarshalNamespace(data []byte, opts ...MarshalOption) (*Namespace, error) 
 	// always skip schema validation on namespaces unmarshal
 	// the namespace is always created by teleport now
 	var namespace Namespace
-	if err := utils.FastUnmarshal(data, &namespace); err != nil {
+	if err := FastUnmarshal(data, &namespace); err != nil {
 		return nil, trace.BadParameter(err.Error())
 	}
