+++ api/types/plugin_data.go
@@ -20,8 +20,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/gravitational/teleport/lib/utils"
-
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
 )
@@ -290,7 +288,7 @@ func (m *pluginDataMarshaler) MarshalPluginData(data PluginData, opts ...Marshal
 			cp.SetResourceID(0)
 			r = &cp
 		}
-		return utils.FastMarshal(r)
+		return FastMarshal(r)
 	default:
 		return nil, trace.BadParameter("unrecognized plugin data type: %T", data)
 	}
@@ -303,11 +301,11 @@ func (m *pluginDataMarshaler) UnmarshalPluginData(raw []byte, opts ...MarshalOpt
 	}
 	var data PluginDataV3
 	if cfg.SkipValidation {
-		if err := utils.FastUnmarshal(raw, &data); err != nil {
+		if err := FastUnmarshal(raw, &data); err != nil {
 			return nil, trace.Wrap(err)
 		}
 	} else {
-		if err := utils.UnmarshalWithSchema(GetPluginDataSchema(), &data, raw); err != nil {
+		if err := UnmarshalWithSchema(GetPluginDataSchema(), &data, raw); err != nil {
 			return nil, trace.Wrap(err)
 		}
 	}
