+++ api/types/saml_test.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/fixtures"
-	"github.com/gravitational/teleport/lib/utils"
 
 	"gopkg.in/check.v1"
 	kyaml "k8s.io/apimachinery/pkg/util/yaml"
@@ -35,7 +34,7 @@ var _ = check.Suite(&SAMLSuite{})
 var _ = fmt.Printf
 
 func (s *SAMLSuite) SetUpSuite(c *check.C) {
-	utils.InitLoggerForTests(testing.Verbose())
+	InitLoggerForTests(testing.Verbose())
 }
 
 func (s *SAMLSuite) TestParseFromMetadata(c *check.C) {
