+++ api/types/traits_test.go
@@ -19,7 +19,6 @@ package types
 import (
 	"testing"
 
-	"github.com/gravitational/teleport/lib/utils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -135,6 +134,6 @@ func TestTraitsToRoleMatchers(t *testing.T) {
 
 		// verify that the resulting matches, once deduplicated, are equivalent
 		// to the expected matches.
-		require.ElementsMatch(t, utils.Deduplicate(matches), tt.matches, tt.desc)
+		require.ElementsMatch(t, Deduplicate(matches), tt.matches, tt.desc)
 	}
 }
