+++ api/types/trustedcluster.go
@@ -23,7 +23,6 @@ import (
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/lib/modules"
-	"github.com/gravitational/teleport/lib/utils"
 
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
@@ -302,7 +301,7 @@ func (c *TrustedClusterV2) CanChangeStateTo(t TrustedCluster) error {
 	if c.GetReverseTunnelAddress() != t.GetReverseTunnelAddress() {
 		return immutableFieldErr("tunnel_addr")
 	}
-	if !utils.StringSlicesEqual(c.GetRoles(), t.GetRoles()) {
+	if !StringSlicesEqual(c.GetRoles(), t.GetRoles()) {
 		return immutableFieldErr("roles")
 	}
 	if !c.GetRoleMap().Equals(t.GetRoleMap()) {
@@ -330,7 +329,7 @@ func (r RoleMapping) Equals(o RoleMapping) bool {
 	if r.Remote != o.Remote {
 		return false
 	}
-	if !utils.StringSlicesEqual(r.Local, r.Local) {
+	if !StringSlicesEqual(r.Local, r.Local) {
 		return false
 	}
 	return true
@@ -371,7 +370,7 @@ func (r RoleMap) parse() (map[string][]string, error) {
 		if roleMap.Remote == "" {
 			return nil, trace.BadParameter("missing 'remote' parameter for role_map")
 		}
-		_, err := utils.ReplaceRegexp(roleMap.Remote, "", "")
+		_, err := ReplaceRegexp(roleMap.Remote, "", "")
 		if trace.IsBadParameter(err) {
 			return nil, trace.BadParameter("failed to parse 'remote' parameter for role_map: %v", err.Error())
 		}
@@ -416,7 +415,7 @@ func (r RoleMap) Map(remoteRoles []string) ([]string, error) {
 				continue
 			}
 			for _, replacementRole := range mapping.Local {
-				replacement, err := utils.ReplaceRegexp(expression, replacementRole, remoteRole)
+				replacement, err := ReplaceRegexp(expression, replacementRole, remoteRole)
 				switch {
 				case err == nil:
 					// empty replacement can occur when $2 expand refers
@@ -530,11 +529,11 @@ func (t *teleportTrustedClusterMarshaler) Unmarshal(bytes []byte, opts ...Marsha
 	}
 
 	if cfg.SkipValidation {
-		if err := utils.FastUnmarshal(bytes, &trustedCluster); err != nil {
+		if err := FastUnmarshal(bytes, &trustedCluster); err != nil {
 			return nil, trace.BadParameter(err.Error())
 		}
 	} else {
-		err := utils.UnmarshalWithSchema(GetTrustedClusterSchema(""), &trustedCluster, bytes)
+		err := UnmarshalWithSchema(GetTrustedClusterSchema(""), &trustedCluster, bytes)
 		if err != nil {
 			return nil, trace.BadParameter(err.Error())
 		}
@@ -568,7 +567,7 @@ func (t *teleportTrustedClusterMarshaler) Marshal(c TrustedCluster, opts ...Mars
 			copy.SetResourceID(0)
 			resource = &copy
 		}
-		return utils.FastMarshal(resource)
+		return FastMarshal(resource)
 	default:
 		return nil, trace.BadParameter("unrecognized resource version %T", c)
 	}
