+++ lib/utils/aliases.go
@@ -1,0 +1,44 @@
+/*
+Copyright 2021 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// type utils
+
+// // oidc.go
+// var (
+// 	ParseSessionsURI = types.ParseSessionsURI
+// )
+
+// // slices.go
+// var (
+// 	CopyByteSlice = types.CopyByteSlice
+// CopyByteSlices = types.CopyByteSlices
+// StringSlicesEqual = types.StringSlicesEqual
+// SliceContainsStr = types.SliceContainsStr
+// Deduplicate = types.Deduplicate
+// )
+
+// // strings.go
+// var (
+// 	CopyStrings = types.CopyStrings
+// )
+
+// // json.go
+// var (
+// 	FastUnmarshal = types.FastUnmarshal
+// FastMarshal = types.FastMarshal
+// UnmarshalWithSchema = types.UnmarshalWithSchema
+// ToJSON = types.ToJSON
+// )
